+++ core/src/include/jcanvas/core/jimage.h
@@ -108,7 +108,7 @@ class Image : public std::enable_shared_from_this<Image> {
      * \brief
      *
      */
-    virtual std::shared_ptr<Image> Rotate(double radians, bool resize = true);
+    virtual std::shared_ptr<Image> Rotate(float radians, bool resize = true);
     
     /**
      * \brief
@@ -126,7 +126,7 @@ class Image : public std::enable_shared_from_this<Image> {
      * \brief
      *
      */
-    virtual std::shared_ptr<Image> Blend(double alpha);
+    virtual std::shared_ptr<Image> Blend(float alpha);
 
     /**
      * \brief
